enabling L2GW to work with DVR
In DVR mode no host IP is associated with the L3 Agent because
the router is configured on all the Compute Nodes and on the Network Node.
To overcome this problem we look for L3 Agent that is running on the Network Node,
resolve the hostname of the Network Node to get its IP address and use it to
configure the destination IP needed for neutron port location information.

Closes-Bug: 1463784

Change-Id: I2595c714ede896baa7726ceec793de9a7a29e6b2